+++ src/resources/af_move.h
@@ -19,7 +19,7 @@ typedef struct af_move_t {
     uint8_t extra_string_selector;
     int16_t damage;
     uint8_t raw_damage; // used to store the raw damage, unmodified by pilot stats
-    float stun;
+    int32_t stun;
     str move_string;
     str footer_string;
 #ifdef DEBUGMODE
